+++ src/formats/png.rs
@@ -30,7 +30,7 @@ pub fn read_tags(src: &mut (impl Read + Seek)) -> Result<crate::TagSet> {
         match &chunk_type {
             END_CHUNK => return Ok(crate::TagSet::new()),
             TAG_CHUNK => {
-                let mut bytes = read_bytes!(src, chunk_length as usize);
+                let mut bytes = read_bytes!(src, chunk_length as u64);
 
                 // Verify checksum
                 let checksum = u32::from_be_bytes(read_bytes!(src, 4));
@@ -66,7 +66,7 @@ pub fn write_tags(src: &mut (impl Read + Seek), dest: &mut impl Write, tags: Tag
     let chunk_type = read_bytes!(src, 4);
     dest.write_all(&chunk_length.to_be_bytes())?;
     dest.write_all(&chunk_type)?;
-    dest.write_all(&read_bytes!(src, chunk_length as usize + 4))?;
+    dest.write_all(&read_bytes!(src, chunk_length as u64 + 4))?;
 
     // Encode tags
     let mut tags: Vec<_> = tags.into_iter().collect();
@@ -110,14 +110,14 @@ pub fn write_tags(src: &mut (impl Read + Seek), dest: &mut impl Write, tags: Tag
             END_CHUNK => {
                 dest.write_all(&chunk_length.to_be_bytes())?;
                 dest.write_all(&chunk_type)?;
-                dest.write_all(&read_bytes!(src, chunk_length as usize + 4))?;
+                dest.write_all(&read_bytes!(src, chunk_length as u64 + 4))?;
                 return Ok(());
             }
             // Leave unrelated chunks unchanged
             _ => {
                 dest.write_all(&chunk_length.to_be_bytes())?;
                 dest.write_all(&chunk_type)?;
-                dest.write_all(&read_bytes!(src, chunk_length as usize + 4))?;
+                dest.write_all(&read_bytes!(src, chunk_length as u64 + 4))?;
             }
         }
     }
